update connectivity file -- many more to include
